// src/features/verification/liveness/types.ts
export enum LivenessState {
  INITIALIZING = 'INITIALIZING',
  POSITIONING = 'POSITIONING',
  CHALLENGE_BLINK = 'CHALLENGE_BLINK',
  CHALLENGE_SMILE = 'CHALLENGE_SMILE',
  CHALLENGE_ROTATION = 'CHALLENGE_ROTATION',
  CHALLENGE_PITCH = 'CHALLENGE_PITCH',
  ANALYZING = 'ANALYZING',
  SUCCESS = 'SUCCESS',
  FAILURE = 'FAILURE',
}